🔧 Modify: 선수 기록 상세보기 컴포넌트로 분리
